ボタン作成

+++ Particle.cpp
@@ -141,7 +141,7 @@ void Particle::EmitterUpdate()
                         pParticle->delta.color = (*emitter)->data.deltaColor;
                     }
 
-                    pParticle->life = (DWORD)(*emitter)->data.lifeTime;    //Žc‚èŽõ–½
+                    pParticle->life = (*emitter)->data.lifeTime;    //Žc‚èŽõ–½
                     pParticle->gravity = (*emitter)->data.gravity;  //d—Í
                     pParticle->accel = (*emitter)->data.accel;      //‰Á‘¬
